Refactor npm resolution
